requirements.txt
pycryptodomex==3.15.0
prettytable==3.3.0
lz4==4.0.1
pytest==7.1.1
pytest-cov==3.0.0
pytest-benchmark==3.4.1
